+++ memtable-core/src/impls/row.rs
@@ -91,7 +91,7 @@ impl<T: Default, const ROW: usize> Table for FixedRowTable<T, ROW> {
         self.col_cnt
     }
 
-    fn get_cell(&self, row: usize, col: usize) -> Option<&Self::Data> {
+    fn cell(&self, row: usize, col: usize) -> Option<&Self::Data> {
         if row < self.row_cnt && col < self.col_cnt {
             Some(&self.cells[row][col])
         } else {
@@ -99,7 +99,7 @@ impl<T: Default, const ROW: usize> Table for FixedRowTable<T, ROW> {
         }
     }
 
-    fn get_mut_cell(&mut self, row: usize, col: usize) -> Option<&mut Self::Data> {
+    fn mut_cell(&mut self, row: usize, col: usize) -> Option<&mut Self::Data> {
         if row < self.row_cnt && col < self.col_cnt {
             Some(&mut self.cells[row][col])
         } else {
@@ -161,7 +161,7 @@ impl<T: Default, const ROW: usize> Table for FixedRowTable<T, ROW> {
     ///
     /// Note that this does **not** remove any cells from the table in their
     /// old positions. Instead, this updates the virtual space within the
-    /// table that is made available for methods like [`Table::get_cell`].
+    /// table that is made available for methods like [`Table::cell`].
     ///
     /// If you want to remove the cells that are no longer within capacity,
     /// call [`Self::truncate`], which will reset them to their default value.
@@ -173,7 +173,7 @@ impl<T: Default, const ROW: usize> Table for FixedRowTable<T, ROW> {
     ///
     /// Note that this does **not** remove any cells from the table in their
     /// old positions. Instead, this updates the virtual space within the
-    /// table that is made available for methods like [`Table::get_cell`].
+    /// table that is made available for methods like [`Table::cell`].
     ///
     /// If you want to remove the cells that are no longer within capacity,
     /// call [`Self::truncate`], which will reset them to their default value.
@@ -286,8 +286,7 @@ impl<T: Default, const ROW: usize> Index<(usize, usize)> for FixedRowTable<T, RO
     /// Indexes into a table by a specific row and column, returning a
     /// reference to the cell if it exists, otherwise panicking
     fn index(&self, (row, col): (usize, usize)) -> &Self::Output {
-        self.get_cell(row, col)
-            .expect("Row/Column index out of range")
+        self.cell(row, col).expect("Row/Column index out of range")
     }
 }
 
@@ -295,7 +294,7 @@ impl<T: Default, const ROW: usize> IndexMut<(usize, usize)> for FixedRowTable<T,
     /// Indexes into a table by a specific row and column, returning a mutable
     /// reference to the cell if it exists, otherwise panicking
     fn index_mut(&mut self, (row, col): (usize, usize)) -> &mut Self::Output {
-        self.get_mut_cell(row, col)
+        self.mut_cell(row, col)
             .expect("Row/Column index out of range")
     }
 }
@@ -339,68 +338,68 @@ mod tests {
     }
 
     #[test]
-    fn get_cell_should_return_ref_to_cell_at_location() {
+    fn cell_should_return_ref_to_cell_at_location() {
         // Sets capacity to that of the 2D array provided
         let table = FixedRowTable::from([vec!["a", "b"], vec!["c", "d"]]);
-        assert_eq!(table.get_cell(0, 0).as_deref(), Some(&"a"));
-        assert_eq!(table.get_cell(0, 1).as_deref(), Some(&"b"));
-        assert_eq!(table.get_cell(1, 0).as_deref(), Some(&"c"));
-        assert_eq!(table.get_cell(1, 1).as_deref(), Some(&"d"));
-        assert_eq!(table.get_cell(1, 2), None);
+        assert_eq!(table.cell(0, 0).as_deref(), Some(&"a"));
+        assert_eq!(table.cell(0, 1).as_deref(), Some(&"b"));
+        assert_eq!(table.cell(1, 0).as_deref(), Some(&"c"));
+        assert_eq!(table.cell(1, 1).as_deref(), Some(&"d"));
+        assert_eq!(table.cell(1, 2), None);
     }
 
     #[test]
-    fn get_cell_should_respect_virtual_boundaries() {
+    fn cell_should_respect_virtual_boundaries() {
         // Sets capacity to that of the 2D array provided
         let mut table = FixedRowTable::from([vec!["a", "b"], vec!["c", "d"]]);
         assert_eq!(table.row_cnt(), 2);
         assert_eq!(table.col_cnt(), 2);
 
-        // If we change the capacity to be smaller, get_cell should respect that
+        // If we change the capacity to be smaller, cell should respect that
         table.set_row_capacity(1);
         table.set_column_capacity(1);
-        assert_eq!(table.get_cell(0, 0).as_deref(), Some(&"a"));
-        assert_eq!(table.get_cell(0, 1).as_deref(), None);
-        assert_eq!(table.get_cell(1, 0).as_deref(), None);
-        assert_eq!(table.get_cell(1, 1).as_deref(), None);
+        assert_eq!(table.cell(0, 0).as_deref(), Some(&"a"));
+        assert_eq!(table.cell(0, 1).as_deref(), None);
+        assert_eq!(table.cell(1, 0).as_deref(), None);
+        assert_eq!(table.cell(1, 1).as_deref(), None);
 
         // Capacity changes don't actually overwrite anything
         table.set_row_capacity(2);
         table.set_column_capacity(2);
-        assert_eq!(table.get_cell(0, 0).as_deref(), Some(&"a"));
-        assert_eq!(table.get_cell(0, 1).as_deref(), Some(&"b"));
-        assert_eq!(table.get_cell(1, 0).as_deref(), Some(&"c"));
-        assert_eq!(table.get_cell(1, 1).as_deref(), Some(&"d"));
+        assert_eq!(table.cell(0, 0).as_deref(), Some(&"a"));
+        assert_eq!(table.cell(0, 1).as_deref(), Some(&"b"));
+        assert_eq!(table.cell(1, 0).as_deref(), Some(&"c"));
+        assert_eq!(table.cell(1, 1).as_deref(), Some(&"d"));
     }
 
     #[test]
-    fn get_mut_cell_should_return_mut_ref_to_cell_at_location() {
+    fn mut_cell_should_return_mut_ref_to_cell_at_location() {
         let mut table = FixedRowTable::from([vec!["a", "b"], vec!["c", "d"]]);
-        *table.get_mut_cell(0, 0).unwrap() = "e";
-        *table.get_mut_cell(0, 1).unwrap() = "f";
-        *table.get_mut_cell(1, 0).unwrap() = "g";
-        *table.get_mut_cell(1, 1).unwrap() = "h";
-        assert_eq!(table.get_mut_cell(2, 0), None);
+        *table.mut_cell(0, 0).unwrap() = "e";
+        *table.mut_cell(0, 1).unwrap() = "f";
+        *table.mut_cell(1, 0).unwrap() = "g";
+        *table.mut_cell(1, 1).unwrap() = "h";
+        assert_eq!(table.mut_cell(2, 0), None);
 
-        assert_eq!(table.get_cell(0, 0).as_deref(), Some(&"e"));
-        assert_eq!(table.get_cell(0, 1).as_deref(), Some(&"f"));
-        assert_eq!(table.get_cell(1, 0).as_deref(), Some(&"g"));
-        assert_eq!(table.get_cell(1, 1).as_deref(), Some(&"h"));
+        assert_eq!(table.cell(0, 0).as_deref(), Some(&"e"));
+        assert_eq!(table.cell(0, 1).as_deref(), Some(&"f"));
+        assert_eq!(table.cell(1, 0).as_deref(), Some(&"g"));
+        assert_eq!(table.cell(1, 1).as_deref(), Some(&"h"));
     }
 
     #[test]
-    fn get_mut_cell_should_respect_virtual_boundaries() {
+    fn mut_cell_should_respect_virtual_boundaries() {
         let mut table = FixedRowTable::from([vec!["a", "b"], vec!["c", "d"]]);
         assert_eq!(table.row_cnt(), 2);
         assert_eq!(table.col_cnt(), 2);
 
-        // If we change the capacity to be smaller, get_mut_cell should respect that
+        // If we change the capacity to be smaller, mut_cell should respect that
         table.set_row_capacity(1);
         table.set_column_capacity(1);
-        assert!(table.get_mut_cell(0, 0).is_some());
-        assert!(table.get_mut_cell(0, 1).is_none());
-        assert!(table.get_mut_cell(1, 0).is_none());
-        assert!(table.get_mut_cell(1, 1).is_none());
+        assert!(table.mut_cell(0, 0).is_some());
+        assert!(table.mut_cell(0, 1).is_none());
+        assert!(table.mut_cell(1, 0).is_none());
+        assert!(table.mut_cell(1, 1).is_none());
     }
 
     #[test]
@@ -409,7 +408,7 @@ mod tests {
 
         assert_eq!(table.insert_cell(0, 0, 123), None);
         assert_eq!(table.insert_cell(0, 0, 999), Some(123));
-        assert_eq!(table.get_cell(0, 0).as_deref(), Some(&999))
+        assert_eq!(table.cell(0, 0).as_deref(), Some(&999))
     }
 
     #[test]
@@ -420,7 +419,7 @@ mod tests {
 
         // Can still create and expand within row in within
         assert_eq!(table.insert_cell(0, 1, 123), None);
-        assert_eq!(table.get_cell(0, 1), Some(&123));
+        assert_eq!(table.cell(0, 1), Some(&123));
 
         // Creating anything outside of the single row will yield nothing
         assert_eq!(table.insert_cell(1, 0, 123), None);