+++ memtable-macros/tests/derive/table/dynamic.rs
@@ -268,12 +268,12 @@ fn should_support_retrieving_typed_columns() {
     table.push_row((false, 123));
     table.push_row((true, 999));
 
-    let mut column = table.column_field1();
+    let mut column = table.field1_column();
     assert_eq!(column.next(), Some(&false));
     assert_eq!(column.next(), Some(&true));
     assert!(column.next().is_none());
 
-    let mut column = table.column_field2();
+    let mut column = table.field2_column();
     assert_eq!(column.next(), Some(&123));
     assert_eq!(column.next(), Some(&999));
     assert!(column.next().is_none());
@@ -285,7 +285,7 @@ fn should_support_converting_into_typed_columns() {
     table.push_row((false, 123));
     table.push_row((true, 999));
 
-    let mut column = table.into_column_field1();
+    let mut column = table.into_field1_column();
     assert_eq!(column.next(), Some(false));
     assert_eq!(column.next(), Some(true));
     assert!(column.next().is_none());
@@ -294,7 +294,7 @@ fn should_support_converting_into_typed_columns() {
     table.push_row((false, 123));
     table.push_row((true, 999));
 
-    let mut column = table.into_column_field2();
+    let mut column = table.into_field2_column();
     assert_eq!(column.next(), Some(123));
     assert_eq!(column.next(), Some(999));
     assert!(column.next().is_none());
@@ -306,28 +306,28 @@ fn should_support_replacing_individual_cells() {
     table.push_row((false, 123));
     table.push_row((true, 999));
 
-    assert_eq!(table.replace_cell_field1(0, true), Some(false));
+    assert_eq!(table.replace_field1(0, true), Some(false));
     {
-        let mut column = table.column_field1();
+        let mut column = table.field1_column();
         assert_eq!(column.next(), Some(&true));
         assert_eq!(column.next(), Some(&true));
         assert!(column.next().is_none());
     }
 
-    assert_eq!(table.replace_cell_field2(1, 0usize), Some(999));
+    assert_eq!(table.replace_field2(1, 0usize), Some(999));
     {
-        let mut column = table.column_field2();
+        let mut column = table.field2_column();
         assert_eq!(column.next(), Some(&123));
         assert_eq!(column.next(), Some(&0));
         assert!(column.next().is_none());
     }
 
-    assert_eq!(table.replace_cell_field2(2, 999usize), None);
+    assert_eq!(table.replace_field2(2, 999usize), None);
     {
         assert_eq!(table.row_cnt(), 2);
         assert_eq!(table.col_cnt(), 2);
 
-        let mut column = table.column_field2();
+        let mut column = table.field2_column();
         assert_eq!(column.next(), Some(&123));
         assert_eq!(column.next(), Some(&0));
         assert!(column.next().is_none());
@@ -340,13 +340,13 @@ fn should_support_retrieving_individual_cells() {
     table.push_row((false, 123));
     table.push_row((true, 999));
 
-    assert_eq!(table.get_cell_field1(0), Some(&false));
-    assert_eq!(table.get_cell_field1(1), Some(&true));
-    assert_eq!(table.get_cell_field1(2), None);
+    assert_eq!(table.field1(0), Some(&false));
+    assert_eq!(table.field1(1), Some(&true));
+    assert_eq!(table.field1(2), None);
 
-    assert_eq!(table.get_cell_field2(0), Some(&123));
-    assert_eq!(table.get_cell_field2(1), Some(&999));
-    assert_eq!(table.get_cell_field2(2), None);
+    assert_eq!(table.field2(0), Some(&123));
+    assert_eq!(table.field2(1), Some(&999));
+    assert_eq!(table.field2(2), None);
 }
 
 #[test]
@@ -355,21 +355,21 @@ fn should_support_mutating_individual_cells() {
     table.push_row((false, 123));
     table.push_row((true, 999));
 
-    *table.get_mut_cell_field1(0).unwrap() = true;
-    *table.get_mut_cell_field1(1).unwrap() = false;
-    assert!(table.get_mut_cell_field1(2).is_none());
+    *table.mut_field1(0).unwrap() = true;
+    *table.mut_field1(1).unwrap() = false;
+    assert!(table.mut_field1(2).is_none());
 
-    *table.get_mut_cell_field2(0).unwrap() = 999;
-    *table.get_mut_cell_field2(1).unwrap() = 123;
-    assert!(table.get_mut_cell_field2(2).is_none());
+    *table.mut_field2(0).unwrap() = 999;
+    *table.mut_field2(1).unwrap() = 123;
+    assert!(table.mut_field2(2).is_none());
 
-    assert_eq!(table.get_cell_field1(0), Some(&true));
-    assert_eq!(table.get_cell_field1(1), Some(&false));
-    assert_eq!(table.get_cell_field1(2), None);
+    assert_eq!(table.field1(0), Some(&true));
+    assert_eq!(table.field1(1), Some(&false));
+    assert_eq!(table.field1(2), None);
 
-    assert_eq!(table.get_cell_field2(0), Some(&999));
-    assert_eq!(table.get_cell_field2(1), Some(&123));
-    assert_eq!(table.get_cell_field2(2), None);
+    assert_eq!(table.field2(0), Some(&999));
+    assert_eq!(table.field2(1), Some(&123));
+    assert_eq!(table.field2(2), None);
 }
 
 #[test]